+++ CHANGELOG.md
@@ -4,6 +4,27 @@ All notable changes to the **TriliumNext-Powershell-Module** will be documented 
 
 ---
 
+## [0.11.0] - 2025-06-13
+
+### Added
+
+* Added function `Get-TriliumNoteAttachment` - Retrieves attachments for a specific note. (Undocumented API)
+* Added `New-TriliumNoteFile` - The File note type can be used to attach various external files such as images, videos or PDF documents.
+
+### Changed
+
+* Changed ETAPI authentication header to use 'Bearer' in `Connect-TriliumAuth`.
+
+### Fixed
+
+* Fixed code block rendering when using `-Markdown` with `New-TriliumNote`.
+
+### Removed
+
+* Removed function `Get-TriliumAttachmentID`, replaced with `Get-TriliumNoteAttachment`
+
+---
+
 ## [0.10.0] - 2025-06-11
 
 ### Added
@@ -13,6 +34,7 @@ All notable changes to the **TriliumNext-Powershell-Module** will be documented 
 * Added function `Get-TriliumAttachmentID` - Parses all attachment IDs from a note's contents (no direct API equivalent).
 * Added parameter `-Math` (used with `-Markdown`) to convert math markdown for rendering.
 * Added function `Create-TriliumAttribute`.
+* Added type `file` to `New-TriliumNote`  for file type notes.
 
 ---